# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chat application with async translation. This API enables users to create chats, send messages, and perform translations asynchronously.",
        "title": "Translatify API",
        "termsOfService": "http://translatify.io/terms/",
        "contact": {
            "name": "danyil.butov",
            "url": "https://github.com/HJyup"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/chats": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new chat between two users.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Create Chat",
                "parameters": [
                    {
                        "description": "Chat information",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.CreateChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chats/{chatId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific chat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Get Chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat details",
                        "schema": {
                            "$ref": "#/definitions/api.GetChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chats/{chatId}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of messages for a specific chat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "List Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Since timestamp (Unix epoch in seconds)",
                        "name": "sinceTimestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of messages to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token for pagination",
                        "name": "pageToken",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "schema": {
                            "$ref": "#/definitions/api.ListMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a message in a chat.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Send Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message information",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chats/{chatId}/messages/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Open a websocket connection to stream messages for a specific chat.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Stream Messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Chat": {
            "type": "object",
            "properties": {
                "Chat_id": {
                    "description": "Unique identifier for the Chat.",
                    "type": "string"
                },
                "created_at": {
                    "description": "Unix timestamp when the Chat was created.",
                    "type": "integer"
                },
                "source_language": {
                    "description": "The language code of the original content (e.g., \"en\" for English).",
                    "type": "string"
                },
                "target_language": {
                    "description": "The target language code for translation (e.g., \"es\" for Spanish).",
                    "type": "string"
                },
                "username_a": {
                    "description": "One participant in the Chat.",
                    "type": "string"
                },
                "username_b": {
                    "description": "The other participant in the Chat.",
                    "type": "string"
                }
            }
        },
        "api.ChatMessage": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "description": "Identifier of the Chat to which this message belongs.",
                    "type": "string"
                },
                "content": {
                    "description": "The original message content.",
                    "type": "string"
                },
                "message_id": {
                    "description": "Unique identifier for the message.",
                    "type": "string"
                },
                "receiver_username": {
                    "description": "The receiver's user ID.",
                    "type": "string"
                },
                "sender_username": {
                    "description": "The sender's user ID.",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix timestamp when the message was created.",
                    "type": "integer"
                },
                "translated_content": {
                    "description": "The translated content (if applicable).",
                    "type": "string"
                }
            }
        },
        "api.CreateChatResponse": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.GetChatResponse": {
            "type": "object",
            "properties": {
                "chat": {
                    "$ref": "#/definitions/api.Chat"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ListMessagesResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "messages": {
                    "description": "The list of messages in the Chat.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ChatMessage"
                    }
                },
                "next_page_token": {
                    "description": "A token that can be used to retrieve the next page of results.",
                    "type": "string"
                }
            }
        },
...
        "api.SendMessageResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message_id": {
                    "description": "Unique identifier for the message.",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.CreateChatRequest": {
            "type": "object",
            "properties": {
                "sourceLanguage": {
                    "type": "string"
                },
                "targetLanguage": {
                    "type": "string"
                },
                "userNameB": {
                    "type": "string"
                },
                "usernameA": {
                    "type": "string"
                }
            }
        },
...
        "models.SendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "fromUsername": {
                    "type": "string"
                },
                "toUsername": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Provide your token with `Bearer \u003ctoken\u003e` format.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}